Partial fix for #595869 ": in URL after ; parsed as list item" Also put sensible names on many variables in doBlockLevels() and added some comments.
This partial fix requires there to be a space prior to the post-; : to get
parsed as the definition:
; def title : def text

thus it doesn't trigger on a URL or namesapce properly written:
; def title [http://link] : def text

However the regexp will fail and break things if there's an actual space
before a : in the link, which is possible in piped text etc. It may be
desirable to deal with this at the tokenizer level, to keep it clear
of links, tags, etc.