Removing the additional protection of the system messages in maintenance/InitialiseMessages.inc. The system messages are already protected using the permission "editinterface", which protects the whole MediaWiki namespace, and they require no additional protection. If we want that regular users will be able to edit the messages, for example, we are currently required to grant them the both permissions "editinterface" and "protect".
From now on, after adding new messages to MediaWiki namespace using the script, the protection level is not set, and the messages are editable only by the users which have the permission "editinterface", regardless the permission "protect" (which should be used for other purposes - protecting pages, and editing protected pages - not for editing system messages). This will separate the permissions, and won't require us to grant the permission "protect" in order to make the "editinterface" permission functional.