* (bug 13880) "Printable" link in Modern skin now formats as print mode
Problem was that the Modern skin uses a different way of setting up the print stylesheet from the other skins.

Others:
* always include main stylesheet
* override that stuff with print stylesheet for print media or ?printable=yes

Modern:
* include main stylesheet only for screen & projection media only
* add some additional print styles for print media or ?printable=yes

The screen stylesheet was still being used when ?printable=yes is set, but would be hidden when actually printing. I've changed it to now exclude the screen stylesheets when ?printable=yes is switched in, so it looks the same on screen as when actually printing.