* (bug 26498) allow LinksUpdate with API
Original patch by Umherirrender. Tweaked to take account of User->pingLimiter() (ie only update links if not limited)

Possible TODO: Error if we are limited? $this->dieUsageMsg( array( 'actionthrottledtext' ) ); seems a bit harsh

Maybe just have a linkupdate="$this->parseMsg( array( 'actionthrottledtext' ) )" in the result? Or something. Definately notification of error to some respect is needed. Just dieing and stopping the whole process is OTT?