Someone on irc had a very screwed up DB, and he reported that the updater gave a fatal error when trying to update (appearently did not have a recentchanges table, and one of the db functions returned false, which the updater did not expect). Add check for that since I figure an exception is much better than a fatal error.
The relavent error this potentially stops is:
Fatal error: Call to a member function isMultipleKey() on a non-object in /../includes/installer/MysqlUpdater.php on line 249"

Not sure if the updater should do something better than an exception, but it seems like
a much better situation than the fatal.

In practise, I think it is rare that this situation would arise.